+++ start/src/App.vue
@@ -3,19 +3,11 @@
 </script>
 
 <template>
-  <f-alert type="default" close>
-    <template #prefixIcon>
-      <f-icon-cross />
-    </template>
-    默认提示信息
-  </f-alert>
-  <f-alert type="primary" close>
-    <template #prefixIcon>
-      <f-icon-cross />
-    </template>
-    主要提示信息
-  </f-alert>
-  <f-alert type="success" close>成功提示信息</f-alert>
-  <f-alert type="danger" close>危险提示信息</f-alert>
-  <f-alert type="warning" close>警告提示信息</f-alert>
+  <f-button type="default">
+    默认按钮
+  </f-button>
+  <f-button type="primary" loading>主要按钮</f-button>
+  <f-button type="success" loading>成功按钮</f-button>
+  <f-button type="danger">危险按钮</f-button>
+  <f-button type="warning">警告按钮</f-button>
 </template>